+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -13,6 +13,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ### Improvements
 
 - Print logs in reverse order (#3260)
+- Allow indicating minimum memory needed by plugin (#3262)
 
 ## 16.0.2 - 15/01/2026