+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to `gatographql` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## 16.1.0 - DATE
+## 16.1.0 - 13/02/2026
 
 ### Added